feat: support video in html block
