+++ .eslintignore
@@ -3,3 +3,5 @@ test/unit/*.js
 test/e2e/*.js
 src/renderer/assets/symbolIcon/index.js
 src/muya/lib/assets/libs/*.js
+
+src/muya/lib/utils/dompurify.js
