+++ src/muya/lib/parser/render/renderBlock/renderLeafBlock.js
@@ -1,9 +1,9 @@
 import katex from 'katex'
 import mermaid from 'mermaid'
 import prism, { loadedCache } from '../../../prism/'
-import { CLASS_OR_ID, DEVICE_MEMORY, PREVIEW_DOMPURIFY_CONFIG, HAS_TEXT_BLOCK_REG } from '../../../config'
+import { CLASS_OR_ID, DEVICE_MEMORY, HAS_TEXT_BLOCK_REG } from '../../../config'
 import { tokenizer } from '../../'
-import { snakeToCamel, sanitize, escapeHtml, getLongUniqueId, getImageInfo } from '../../../utils'
+import { snakeToCamel, escapeHtml, getLongUniqueId } from '../../../utils'
 import { h, htmlToVNode } from '../snabbdom'
 
 // todo@jocs any better solutions?
@@ -107,22 +107,10 @@ export default function renderLeafBlock (block, cursor, activeBlocks, selectedBl
     switch (functionType) {
       case 'preview': {
         selector += `.${CLASS_OR_ID['AG_HTML_PREVIEW']}`
-        const htmlContent = sanitize(code, PREVIEW_DOMPURIFY_CONFIG)
-        // handle empty html bock
-        if (/^<([a-z][a-z\d]*)[^>]*?>(\s*)<\/\1>$/.test(htmlContent.trim())) {
-          children = htmlToVNode('<div class="ag-empty">&lt;Empty HTML Block&gt;</div>')
-        } else {
-          const parser = new DOMParser()
-          const doc = parser.parseFromString(htmlContent, 'text/html')
-          const imgs = doc.documentElement.querySelectorAll('img')
-          for (const img of imgs) {
-            const src = img.getAttribute('src')
-            const imageInfo = getImageInfo(src)
-            img.setAttribute('src', imageInfo.src)
-          }
-
-          children = htmlToVNode(doc.documentElement.querySelector('body').innerHTML)
-        }
+        this.htmlCache.set(`#${block.key}`, {
+          code
+        })
+        children = ''
         break
       }
       case 'multiplemath': {
